fix: remove r0 from challenges list
r0 is the UniSkip challenge and should NOT be in the sumcheck challenges list.
The challenges list should only contain [r_stream, r_1, ..., r_n] which are
the streaming round and cycle round challenges.

This fixes the r_cycle computation in proof converter which was off-by-one
because it was skipping what it thought was r_stream but was actually r0.